Extract private ES2K functions
- Extracted private ES2K functions from ovsp4rt.cc and moved them
  to ovsp4rt_es2k_private.cc. Created ovsp4rt_es2k_private.h to
  provide function prototypes for the new module.
